Fix TableView UITk init and attach to Panel. Split TableView in different file (UITk and IMGUI)
